refactor functions into utils
